refactor(agent): split output-budget and calibration into budget.go
compact.go exceeded the 800-line repolint ceiling (974 lines); the
shared-window budget clip, prompt-token estimation, resume gate, and
tokPerChar calibration move to internal/agent/budget.go. Comments are
trimmed to the enforced lengths. No behavior change.
